admin_panel/requirements.admin.txt
# 📋 LegalRAG Admin Panel Dependencies

# Core Streamlit
streamlit==1.28.2
streamlit-authenticator==0.2.3

# API Requests
requests==2.31.0
httpx==0.25.2

# JWT Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# Data Processing
pandas==2.1.4
numpy==1.24.4

# Excel and Spreadsheet Support (для новой функциональности)
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0

# Charts and Visualizations  
plotly==5.17.0
altair==5.1.2

# File Operations
python-multipart==0.0.6

# PDF Processing (для анализа ориентации PDF)
PyPDF2==3.0.1

# Background Tasks (встроенный threading вместо threading2)

# Time and Date
python-dateutil==2.8.2

# Database (same as main app) - обновленные версии для v2.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Async support for user management
aioredis==2.0.1

# Environment
python-dotenv==1.0.0

# Utilities
pathvalidate==3.2.0
humanize==4.8.0
psutil==6.0.0

# WebSocket for real-time logs
websockets==12.0

# Testing and E2E (для Playwright тестов)
playwright==1.40.0
pytest-playwright==0.4.3
pytest-xvfb==3.0.0